server/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-RESTful==0.3.10
Flask-CORS==4.0.0
SQLAlchemy-Serializer==1.4.1
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7